requirements.txt
selenium
pytest
allure-pytest
webdriver-manager
pytest-soft-asserts


#to run the requirements.txt run the below command
#requiremnts.txt are the dependiencies for the project
#pip install -r requirements.txt